remove Cargo.lock to improve ease-of-use
When `Cargo.lock` is present, all dependencies need to be present in the versions noted within.
As `node` needs to be a sibling of `wc-demo` though, it is likely the user will
download the latest version, which might not match the required one.

Removing Cargo.lock allows all versions within the boundary specified in
`Cargo.toml`, and increases the chance the build will succeed.